No need for signs

+++ 001/solution_001_rs.rs
@@ -1,4 +1,4 @@
-fn solution(upto: i32) -> i32 {
+fn solution(upto: u64) -> u64 {
     (1..upto).filter(|i| i % 3 == 0 || i % 5 == 0).sum()
 }
